+++ webpack.config.cjs
@@ -76,12 +76,12 @@ module.exports = {
 				// share map.
 				react: {
 					singleton: true,
-					requiredVersion: ">=19.2.0 <20.0.0",
+					requiredVersion: "^19.2.0",
 					import: false,
 				},
 				"react-dom": {
 					singleton: true,
-					requiredVersion: ">=19.2.0 <20.0.0",
+					requiredVersion: "^19.2.0",
 					import: false,
 				},
 			},